backend/requirements.txt
flask
pymongo
flask-cors
python-dotenv
flask-bcrypt
flask-jwt-extended
